requirements.txt
python-dotenv==1.21.1

tapo==0.8.7
uvicorn==0.39.0
fastapi==0.128.0
